+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 1.1 branch
 ### 1.1.7
 * added `is_true_or_fail()` function;
-* added `create_file()` and `create_tmp_file()` global functions;
+* added `create_file()` and `create_tmp_file()` global functions and
+    `safe_create_file()` and `safe_create_tmp_file()` safe functions;
 * `file_exists_or_fail()`, `is_dir_or_fail()`, `is_readable_or_fail()` and
     `is_writable_or_fail()` functions are now deprecated and they will be
     removed from the `1.2.0` version. Use instead `is_true_or_fail()`.